refactor(team): extract runtime-agents + agent-lessons helpers from index
Phase A of the team/index split (WS6). The 2,312-line server entry sheds two
cohesive, self-contained helper groups into lib/ modules (matching the existing
lib/ idiom — session-reader, build-graph, health-checks):

- lib/runtime-agents.ts — the ~270-line runtime-adapter agent wrapper layer
  (agentToMeta, list/get/create/remove/update agents, allowlists, subagent
  perms, skills, memory files, heartbeat, identity-markdown synth/parse). Every
  function takes the AgentRuntimeAdapter explicitly or is a pure transform — no
  module-level context coupling.
- lib/agent-lessons.ts — agent-package lesson path parsing + fs reads for the
  bakin_agent-lessons search content type (parse path, file->id, key->path,
  frontmatter, file->doc, reindex generator).

index.ts: 2,312 -> 1,914 lines. Pure relocation — populateTeamRoutes and
activate() are untouched (route handlers already take ctx as a parameter; they
now import the helpers instead of closing over module-scope functions).

Deferred to a follow-up: extracting the ~970-line populateTeamRoutes into
lib/routes/* (the bulk) and the status-resolution group.

Verified: typecheck + lint clean; team suite 183/0; full suite 5123/0; binary
build; isolated-home boot smoke (Team activates cleanly through the refactored
path — "Plugin loaded: Team v1.0.0", team onReady fires).
