+++ plugins/team/index.ts
@@ -15,25 +15,22 @@ import {
   openSync,
   readFileSync,
   readSync,
-  readdirSync,
-  statSync,
   unlinkSync,
   writeFileSync,
 } from 'fs'
-import { basename, dirname, join, relative } from 'path'
+import { dirname, join } from 'path'
 import type { BakinPlugin, PluginContext } from '@bakin/core/plugin-types'
 import { defineRoute, definePlugin } from '@bakin/core/routing'
 import type { PluginContextLite } from '@bakin/core/routing'
 import { createLogger } from '../../src/core/logger'
 import { readHeartbeats } from '../../src/lib/content-files'
 import { getContentDir, getBakinPaths } from '../../packages/core/src/content-dir'
-import { resolveAgentAvatar, serveAvatar, detectImageExtension } from '@bakin/core/agents/avatar'
+import { serveAvatar, detectImageExtension } from '@bakin/core/agents/avatar'
 import { removeInstalledBy } from '@bakin/core/agent-packages/markers'
 import {
   readPluginSettings as readStoredPluginSettings,
   writePluginSettings as writeStoredPluginSettings,
 } from '@bakin/core/plugins/settings-store'
-import { parseLessonFrontmatter as parseLessonFm } from '@bakin/core/format/frontmatter'
 import { startAgent, stopAgent } from '../../src/lib/agents'
 import { getSettings, resetSettingsCache } from '../../src/core/settings'
 import { syncConfig as syncMcporter } from '../../src/core/mcporter'
@@ -43,18 +40,38 @@ import { appendAudit } from '../../src/core/audit'
 import { getAllAgentUsage } from '../../src/core/agent-usage'
 import { getStatsByMs } from '../../src/core/usage'
 import { retrieveAgentPackageLessons } from '../../src/core/agent-packages/lesson-retrieval'
-import { getRuntimeMainAgentId, type AgentRuntimeAdapter, type RuntimeAgent } from '@bakin/core/adapters/runtime'
+import { getRuntimeMainAgentId, type AgentRuntimeAdapter } from '@bakin/core/adapters/runtime'
 import { readLatestSessionTranscript } from './lib/session-reader'
 import { agentSyncRepair, checkAgentRoster, checkPersonas, checkAgentSync, personaRepair } from './lib/health-checks'
+import {
+  agentToMeta,
+  listRuntimeAgentMetas,
+  getRuntimeAgentIds,
+  getRuntimeAgentModel,
+  getRuntimeAgentProfile,
+  createRuntimeAgent,
+  addToRuntimeAllowlists,
+  removeFromRuntimeAllowlists,
+  removeRuntimeAgent,
+  updateRuntimeAgentIdentity,
+  setRuntimeSubagentPermissions,
+  listRuntimeSkills,
+  readRuntimeSkillFile,
+  listRuntimeMemoryFiles,
+  readRuntimeMemoryFile,
+  readRuntimeHeartbeatRaw,
+} from './lib/runtime-agents'
+import {
+  agentLessonFileToId,
+  agentLessonKeyToFilePath,
+  agentLessonFileToDoc,
+  agentLessonReindexAll,
+} from './lib/agent-lessons'
 import type {
-  AgentMeta,
-  AgentProfile,
   AgentWithStatus,
   AgentDisplaySettingsMap,
   HeartbeatData,
-  HeartbeatRaw,
   OrgTeam,
-  SkillSummary,
   TeamPluginSettings,
 } from './types'
 
@@ -157,283 +174,6 @@ async function mergeDisplayDefaults(runtime: AgentRuntimeAdapter, overrides: Age
   return result
 }
 
-interface IdentityFields {
-  name?: string
-  emoji?: string
-  role?: string
-  vibe?: string
-  primaryFunction?: string
-  defaultMode?: string
-}
-
-interface CreateAgentInput extends IdentityFields {
-  id: string
-  name: string
-  model?: string
-  soul?: string
-  tools?: string
-}
-
-function metadataString(agent: RuntimeAgent, key: string): string | undefined {
-  const value = agent.metadata?.[key]
-  return typeof value === 'string' && value.length > 0 ? value : undefined
-}
-
-function metadataStringArray(agent: RuntimeAgent, key: string): string[] | null {
-  const value = agent.metadata?.[key]
-  return Array.isArray(value) && value.every((entry) => typeof entry === 'string') ? value : null
-}
-
-function agentToMeta(agent: RuntimeAgent): AgentMeta {
-  const headshot = resolveAgentAvatar(agent.id) ? `/api/plugins/team/${agent.id}/avatar` : ''
-
-  return {
-    id: agent.id,
-    name: agent.name || agent.id,
-    emoji: metadataString(agent, 'emoji') ?? '',
-    role: agent.role ?? metadataString(agent, 'role') ?? '',
-    headshot,
-  }
-}
-
-async function listRuntimeAgentMetas(runtime: AgentRuntimeAdapter): Promise<AgentMeta[]> {
-  return (await runtime.agents.list()).map(agentToMeta)
-}
-
-async function getRuntimeAgentIds(runtime: AgentRuntimeAdapter): Promise<string[]> {
-  return (await runtime.agents.list()).map((agent) => agent.id)
-}
-
-async function getRuntimeAgentModel(runtime: AgentRuntimeAdapter, agent: RuntimeAgent): Promise<string> {
-  if (agent.model) return agent.model
-  const config = await runtime.config.get<{
-    agents?: { defaults?: { model?: { primary?: string } } }
-  }>()
-  return config.agents?.defaults?.model?.primary ?? 'unknown'
-}
-
-async function readRuntimeWorkspaceText(
-  runtime: AgentRuntimeAdapter,
-  agentId: string,
-  path: string,
-): Promise<string | null> {
-  return (await runtime.agents.readWorkspaceFile(agentId, path))?.content ?? null
-}
-
-async function getRuntimeAgentProfile(runtime: AgentRuntimeAdapter, agentId: string): Promise<AgentProfile | null> {
-  const agent = await runtime.agents.get(agentId)
-  if (!agent) return null
-  const meta = agentToMeta(agent)
-  const model = await getRuntimeAgentModel(runtime, agent)
-  return {
-    ...meta,
-    model,
-    workspacePath: metadataString(agent, 'workspacePath') ?? '',
-    soul: await readRuntimeWorkspaceText(runtime, agent.id, 'SOUL.md'),
-    identity: await readRuntimeWorkspaceText(runtime, agent.id, 'IDENTITY.md'),
-    rules: await readRuntimeWorkspaceText(runtime, agent.id, 'AGENTS.md'),
-    tools: await readRuntimeWorkspaceText(runtime, agent.id, 'TOOLS.md'),
-    heartbeatMd: await readRuntimeWorkspaceText(runtime, agent.id, 'HEARTBEAT.md'),
-    subagentPerms: metadataStringArray(agent, 'subagentAllowAgents'),
-  }
-}
-
-function synthesizeIdentityMd(fields: IdentityFields): string {
-  const lines = ['# IDENTITY.md', '']
-  const entries: [string, string | undefined][] = [
-    ['Name', fields.name],
-    ['Role', fields.role],
-    ['Emoji', fields.emoji],
-    ['Vibe', fields.vibe],
-    ['Primary Function', fields.primaryFunction],
-    ['Default Mode', fields.defaultMode],
-  ]
-  for (const [label, value] of entries) {
-    if (value) lines.push(`- **${label}:** ${value}`)
-  }
-  lines.push('')
-  return lines.join('\n')
-}
-
-function parseIdentityMd(content: string): Record<string, string> {
-  const fields: Record<string, string> = {}
-  const regex = /^[-*]\s*\*\*(.+?):\*\*\s*(.+)/gm
-  let match
-  while ((match = regex.exec(content)) !== null) {
-    fields[match[1]] = match[2].trim()
-  }
-  return fields
-}
-
-async function createRuntimeAgent(
-  runtime: AgentRuntimeAdapter,
-  input: CreateAgentInput,
-): Promise<{ id: string; workspace: string }> {
-  const agent = await runtime.agents.create({
-    id: input.id,
-    name: input.name,
-    role: input.role,
-    model: input.model,
-    metadata: {
-      emoji: input.emoji,
-      role: input.role,
-      vibe: input.vibe,
-      primaryFunction: input.primaryFunction,
-      defaultMode: input.defaultMode,
-    },
-  })
-
-  await runtime.agents.writeWorkspaceFile(input.id, {
-    path: 'IDENTITY.md',
-    content: synthesizeIdentityMd(input),
-  })
-  if (input.soul) {
-    await runtime.agents.writeWorkspaceFile(input.id, { path: 'SOUL.md', content: input.soul })
-  }
-  if (input.tools) {
-    await runtime.agents.writeWorkspaceFile(input.id, { path: 'TOOLS.md', content: input.tools })
-  }
-
-  return { id: input.id, workspace: metadataString(agent, 'workspacePath') ?? '' }
-}
-
-async function addToRuntimeAllowlists(
-  runtime: AgentRuntimeAdapter,
-  newAgentId: string,
-  dispatchable: 'all' | 'main' | string[],
-): Promise<void> {
-  const mainAgentId = await getRuntimeMainAgentId(runtime)
-  if (dispatchable === 'main') {
-    await runtime.agents.updateAllowlist(mainAgentId, { add: [newAgentId] })
-    return
-  }
-
-  const agents = await runtime.agents.list()
-  if (dispatchable === 'all') {
-    await Promise.all(
-      agents
-        .filter((agent) => agent.id !== newAgentId)
-        .map((agent) => runtime.agents.updateAllowlist(agent.id, { add: [newAgentId] })),
-    )
-    return
-  }
-
-  const targetIds = new Set(dispatchable)
-  targetIds.add(mainAgentId)
-  targetIds.delete(newAgentId)
-  await Promise.all(Array.from(targetIds).map((agentId) => runtime.agents.updateAllowlist(agentId, { add: [newAgentId] })))
-}
-
-async function removeFromRuntimeAllowlists(runtime: AgentRuntimeAdapter, agentId: string): Promise<void> {
-  const agents = await runtime.agents.list()
-  await Promise.all(agents.map((agent) => runtime.agents.updateAllowlist(agent.id, { remove: [agentId] })))
-}
-
-async function removeRuntimeAgent(runtime: AgentRuntimeAdapter, agentId: string): Promise<boolean> {
-  if (!(await runtime.agents.get(agentId))) return false
-  await runtime.agents.remove(agentId)
-  return true
-}
-
-async function updateRuntimeAgentIdentity(
-  runtime: AgentRuntimeAdapter,
-  agentId: string,
-  fields: IdentityFields & { soul?: string; tools?: string },
-): Promise<string[]> {
-  const agent = await runtime.agents.get(agentId)
-  if (!agent) throw new Error(`Agent "${agentId}" not found in roster`)
-
-  const updated: string[] = []
-  if (fields.name || fields.emoji || fields.role || fields.vibe || fields.primaryFunction || fields.defaultMode) {
-    await runtime.agents.update(agentId, {
-      name: fields.name,
-      role: fields.role,
-      metadata: {
-        ...(agent.metadata ?? {}),
-        ...(fields.emoji ? { emoji: fields.emoji } : {}),
-        ...(fields.role ? { role: fields.role } : {}),
-        ...(fields.vibe ? { vibe: fields.vibe } : {}),
-        ...(fields.primaryFunction ? { primaryFunction: fields.primaryFunction } : {}),
-        ...(fields.defaultMode ? { defaultMode: fields.defaultMode } : {}),
-      },
-    })
-    if (fields.name) updated.push('name')
-    if (fields.emoji) updated.push('emoji')
-  }
-
-  const structuredFields = ['role', 'vibe', 'primaryFunction', 'defaultMode'] as const
-  const hasStructuredUpdate = structuredFields.some((field) => fields[field])
-  if (hasStructuredUpdate || fields.name || fields.emoji) {
-    const existing = await readRuntimeWorkspaceText(runtime, agentId, 'IDENTITY.md')
-    const parsed = existing ? parseIdentityMd(existing) : {}
-    const merged: IdentityFields = {
-      name: fields.name ?? parsed['Name'],
-      emoji: fields.emoji ?? parsed['Emoji'],
-      role: fields.role ?? parsed['Role'],
-      vibe: fields.vibe ?? parsed['Vibe'],
-      primaryFunction: fields.primaryFunction ?? parsed['Primary Function'],
-      defaultMode: fields.defaultMode ?? parsed['Default Mode'],
-    }
-    await runtime.agents.writeWorkspaceFile(agentId, { path: 'IDENTITY.md', content: synthesizeIdentityMd(merged) })
-    for (const field of structuredFields) {
-      if (fields[field]) updated.push(field)
-    }
-  }
-
-  if (fields.soul) {
-    await runtime.agents.writeWorkspaceFile(agentId, { path: 'SOUL.md', content: fields.soul })
-    updated.push('soul')
-  }
-  if (fields.tools) {
-    await runtime.agents.writeWorkspaceFile(agentId, { path: 'TOOLS.md', content: fields.tools })
-    updated.push('tools')
-  }
-
-  return updated
-}
-
-async function setRuntimeSubagentPermissions(
-  runtime: AgentRuntimeAdapter,
-  agentId: string,
-  allowAgents: string[],
-): Promise<void> {
-  if (allowAgents.includes(agentId)) {
-    throw new Error(`Agent "${agentId}" cannot dispatch to itself`)
-  }
-  await runtime.agents.updateAllowlist(agentId, { replace: allowAgents })
-}
-
-async function listRuntimeSkills(runtime: AgentRuntimeAdapter, agentId: string): Promise<SkillSummary[]> {
-  return (await runtime.skills.list(agentId)).map((skill) => ({
-    id: skill.name,
-    name: skill.name,
-    hasSkillMd: typeof skill.metadata?.hasSkillMd === 'boolean'
-      ? skill.metadata.hasSkillMd
-      : Boolean(skill.instructions || skill.path),
-  }))
-}
-
-async function readRuntimeSkillFile(runtime: AgentRuntimeAdapter, agentId: string, skillId: string): Promise<string | null> {
-  return (await runtime.skills.get(skillId, agentId))?.instructions ?? null
-}
-
-async function listRuntimeMemoryFiles(runtime: AgentRuntimeAdapter, agentId: string): Promise<string[]> {
-  return (await runtime.memory.listEntries('workspace-memory', { agentId }))
-    .map((entry) => basename(entry.path ?? entry.id))
-    .sort()
-    .reverse()
-}
-
-async function readRuntimeMemoryFile(runtime: AgentRuntimeAdapter, agentId: string, date: string): Promise<string | null> {
-  const entry = await runtime.memory.getEntry('workspace-memory', date, { agentId })
-  if (entry) return entry.content
-  return readRuntimeWorkspaceText(runtime, agentId, `memory/${date}`)
-}
-
-async function readRuntimeHeartbeatRaw(runtime: AgentRuntimeAdapter, agentId: string): Promise<HeartbeatRaw | null> {
-  const heartbeat = await runtime.agents.readWorkspaceFile(agentId, 'HEARTBEAT.md')
-  return heartbeat ? { content: heartbeat.content, lastUpdated: heartbeat.updatedAt ?? null } : null
-}
 
 // ─── Status Resolution ───────────────────────────────────────────────────────
 
@@ -569,144 +309,6 @@ async function getOrgStructure(runtime: AgentRuntimeAdapter) {
 
 /** Module-level hook for batch-indexing agents — set during activate() */
 let batchIndexAgents: () => Promise<void> = async () => {}
-
-// ─── Agent-lessons indexing helpers ────────────────────────────────────────
-
-/**
- * Decompose a packages/agents/<dir>@<version>/lessons/<lesson>.md path
- * into the parts needed to build a search id + doc. Returns null on a
- * non-conforming path (the watcher passes us anything that matches the
- * glob; we trust the glob for happy-path inputs but defensively handle
- * the edge cases).
- */
-interface LessonFileParts {
-  packageId: string
-  version: string
-  lessonId: string
-  agentId: string
-}
-
-function parseLessonFilePath(rel: string): LessonFileParts | null {
-  // rel example: 'packages/agents/pixel@0.1.0/lessons/style.md'
-  const segments = rel.split('/')
-  if (segments.length < 5) return null
-  if (segments[0] !== 'packages' || segments[1] !== 'agents') return null
-  if (segments[3] !== 'lessons') return null
-  const dirName = segments[2]
-  const lessonFile = segments[4]
-  if (!lessonFile.endsWith('.md')) return null
-  const at = dirName.lastIndexOf('@')
-  if (at === -1) return null
-  const packageId = dirName.slice(0, at)
-  const version = dirName.slice(at + 1)
-  const lessonId = lessonFile.replace(/\.md$/i, '')
-  // For kind:"agent" packages the package id IS the agent id; this matches
-  // the projection convention in src/core/agent-packages/projector.ts.
-  return { packageId, version, agentId: packageId, lessonId }
-}
-
-function agentLessonFileToId(rel: string): string {
-  const parts = parseLessonFilePath(rel)
-  if (!parts) return rel.replace(/[^a-zA-Z0-9_]+/g, '_')
-  return `${parts.packageId}@${parts.version}/${parts.lessonId}`
-}
-
-function agentLessonKeyToFilePath(key: string): string | null {
-  const slash = key.indexOf('/')
-  if (slash === -1) return null
-  const dirPart = key.slice(0, slash)
-  const lesson = key.slice(slash + 1)
-  return join(getContentDir(), 'packages', 'agents', dirPart, 'lessons', `${lesson}.md`)
-}
-
-interface LessonDoc extends Record<string, unknown> {
-  title: string
-  body: string
-  package_id: string
-  agent_id: string
-  lesson_id: string
-  tags: string[]
-  default_enabled: 'true' | 'false'
-  updated_at: string
-}
-
-function parseLessonFrontmatter(raw: string): {
-  title: string
-  body: string
-  tags: string[]
-  defaultEnabled: boolean
-} {
-  const { title, body, tags, defaultEnabled } = parseLessonFm(raw)
-  return { title: title ?? '', body, tags, defaultEnabled }
-}
-
-async function agentLessonFileToDoc(rel: string): Promise<LessonDoc | null> {
-  const parts = parseLessonFilePath(rel)
-  if (!parts) return null
-  const abs = join(getContentDir(), rel)
-  if (!existsSync(abs)) return null
-  let raw: string
-  try {
-    raw = readFileSync(abs, 'utf-8')
-  } catch {
-    return null
-  }
-  const { title, body, tags, defaultEnabled } = parseLessonFrontmatter(raw)
-  return {
-    title: title || parts.lessonId,
-    body,
-    package_id: parts.packageId,
-    agent_id: parts.agentId,
-    lesson_id: parts.lessonId,
-    tags,
-    default_enabled: defaultEnabled ? 'true' : 'false',
-    updated_at: new Date().toISOString(),
-  }
-}
-
-/**
- * Walk every installed agent package's lessons/ dir and yield
- * (key, doc) pairs for the search index reindexer.
- */
-function* agentLessonReindexAll(): Generator<{ key: string; doc: LessonDoc }> {
-  const root = join(getContentDir(), 'packages', 'agents')
-  if (!existsSync(root)) return
-  for (const dirent of readdirSync(root, { withFileTypes: true })) {
-    if (!dirent.isDirectory()) continue
-    const lessonsDir = join(root, dirent.name, 'lessons')
-    if (!existsSync(lessonsDir)) continue
-    let stat
-    try { stat = statSync(lessonsDir) } catch { continue }
-    if (!stat.isDirectory()) continue
-    for (const file of readdirSync(lessonsDir)) {
-      if (!file.endsWith('.md')) continue
-      const rel = relative(getContentDir(), join(lessonsDir, file))
-      const parts = parseLessonFilePath(rel)
-      if (!parts) continue
-      const raw = readFileSync(join(lessonsDir, file), 'utf-8')
-      const { title, body, tags, defaultEnabled } = parseLessonFrontmatter(raw)
-      const key = `${parts.packageId}@${parts.version}/${parts.lessonId}`
-      yield {
-        key,
-        doc: {
-          title: title || parts.lessonId,
-          body,
-          package_id: parts.packageId,
-          agent_id: parts.agentId,
-          lesson_id: parts.lessonId,
-          tags,
-          default_enabled: defaultEnabled ? 'true' : 'false',
-          updated_at: new Date().toISOString(),
-        },
-      }
-    }
-  }
-}
-
-// Reference imports so unused-import linting doesn't fire when the
-// helpers above don't exercise every utility on every code path.
-void basename
-void dirname
 
 // ─── Plugin context handle (set during activate, used by module-scope helpers) ─
