+++ tasks/plan.md
@@ -128,8 +128,16 @@ pre-existing duplicate — WS6 workflows-split dedup territory; noted there.)
   (10 plugins) + host `build.ts` smoke.
 - Remaining WS7: shared dir-walker.
 
-## WS6 — plugin god-files (not started)
+## WS6 — plugin god-files
 
-team/index (2,312), workflows/index (2,146) + lib/runtime (1,633), workflow-canvas-editor (1,803),
-models-page (1,205), health-page (1,155), schedule/index (1,443), tasks/index (1,089), asset-service (835)
-+ test splits. Includes the workflows validator dedup flagged in #510.
+- team/index — ◧ **Phase A done.** Extracted two cohesive lib modules from the 2,312-line index:
+  `lib/runtime-agents.ts` (the ~270-line runtime-adapter agent wrapper layer — every fn takes the adapter
+  explicitly, zero module-ctx coupling) and `lib/agent-lessons.ts` (the agent-package lesson path-parse +
+  fs-read helpers for the `bakin_agent-lessons` content type). index.ts 2,312 → 1,914; no behavior change
+  (pure relocation of param-based/pure helpers; routes + activate untouched). Verified: typecheck/lint/team
+  suite 183-0/full-suite/binary build. DEFERRED (Phase B): extract the ~970-line `populateTeamRoutes` into
+  `lib/routes/*` (the bulk; route handlers take ctx as a param so it's tractable) + the status-resolution
+  group (`staleSettingsCtx`/`resolveAgentStatus`/`getOrgStructure`).
+- Remaining: workflows/index (2,146) + lib/runtime (1,633), workflow-canvas-editor (1,803), models-page
+  (1,205), health-page (1,155), schedule/index (1,443), tasks/index (1,089), asset-service (835) + test
+  splits. Includes the workflows validator dedup flagged in #510.
